requirements/data_scraper.txt
pandas==0.24.1
Flask==1.1.2
flask_restx
mysql-connector-python==8.0.14